add/ favorites and user state connected
